webapi/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
shap>=0.42.0
orjson>=3.9.0
fastf1>=3.3.0
python-multipart>=0.0.6
pyyaml>=6.0.0
